Overhauled the input filtering functionality of Jomres, removing the < x > code that's inserted by our own remove xss code. Whilst it was lovely and safe, it tends to interfere too much with things like tasks and email addresses.
Instead the html purifier functionality has been rewritten to be called as a singleton in a new class. This appears to be much quicker than we did it before because instead of initialising a new class every time, we're reusing the same instance. We've also upgraded html purifier to 4.3.0, instead of the older, obsolete 4.0.