Modified a previously removed switch. Put it back in, but changed how it works.
Previously we'd removed the "Enable currency conversion" option, because for the shopping cart to work the conversion functionality ALWAYS has to be enabled. Some people didn't like this as they only wanted prices output in one currency, nnn.nn instead of nnn.nn (nnn.nn). 
To get around this user conflict we've re-added the switch, but changed what it effects. Instead of using it to switch off the conversion functionality internally, we'll leave the functionality enabled, instead we're simply using it to tell the "output_price" function to not output the price in the nnn.nn (nnn.nn) format.