We've added a complete new global rooms feature that works like business features. This will alow managers to select room features when editing the room, without the need to create them first in the room features tab. The room features tab and functionality is still kept, so managers can still add their own room features too.
Business features feature has been changed, so now we can assign more property types to one feature instead of duplicating it for each property type where we want to show it.

Added a cms specific change to ensure we're getting the SEF setting correctly.

For Joomla 3.1 installations we've removed Jomres native IDN, as J3.1 has this feature now.

In database.php, we've added query caching. It`s not visible when one guest is visiting the website, but when more are doing it at the same time, queries will be executed once and results provided to all, instead of executing the queries for each guest.

Found and resolved an issue on the booking form when rooms list style is enabled. When deselecting rooms, they were not unlocked.

Added Superior as image, same size as stars. This will make it display better. Also add S to property list list with maps, module popup and ngm module output.

Fixed a minor issue where featured listings css class was not added to the next results displayed when live scrolling.

j16000save_custom_field.class.php moved into the Custom Field plugin, where it belongs.

Modified several default language files so that instead of referring to "Room features" we're using language more appropriate to the situation.