Added Font Awesome as an optional css library. Many BS2 Joomla templates exclude glyph icons, or name the icons differently, meaning we can't always rely on a template to provide the required icons, so we'll add our own. Switchable because all BS3 joomla templates so far seem to include FA icon sets by default.
Modified the Booking Engine and Search functionality so that calendars now use calendar icons instead of the old calendar image. Makes styling much easier and looks neater.

The clear dates checkbox can now be added to a search template via {CLEARDATES_CHECKBOX} (and associated checkbox can be removed from the js_calendar_input.html template. This helps with styling.

Output price has been changed to use an icon instead of a link on the price to allow users to see the converted price. This is likely to change, and is just for experimentation at this point.