Improved new code to 1. Simplify & improve gateway validation function and 2 allow a gateway's 00509 script to send a "connected" state that will set the gateway to active.
The Stripe gateway doesn't have a frontend "active" setting that can be pulled from the plugin settings table, because Stripe settings are per manager, not per property. Various pages, however, demand that the active setting be true so that the gateway can be shown as an option. Therefore we'll use the "connected" state to enable the Stripe gateways. This would also work on any other gateway that follows the same pattern.