Added a new setting to Tariffs and Currencies tab which allows property managers to save a basic price that would be shown if no prices can otherwise be determined.
When properties are shown in a list, the system will attempt to find a valid price from configured tariffs based on either today's date, or if dates were used in a search, then based on those dates. If it cannot, then it will show POA (Price on application), meaning that the guest should contact you to get a price. If you want, you can configure a price here that will be shown instead of the POA text. This figure would be a "fallback" price to display if no other price can be determined.