Update fr-FR.php
Hi againVince,

Here is the version 2 of the translation of Jomres, following test on site. I found some syntax errors I had made. I just tested it, it does not return syntax error in the back-end ...

I will now be able to activate the fr-FR.php to re-read all the translations in their context

I was also looking for the files in the following folders that I will translate as soon as possible ...

/administrator/components/com_jomres/language/en-GB
/administrator/language/en-GB
/modules/mod_itpsocialbuttons/language/en-GB
/modules/mod_vina_camera_image_slider/languages
/templates/jr_leotian/language/en-GB

To be continued

Best regards