Added significant update to user handling within Jomres to allow for private and public user data and user profiles.
Users are now able to add "About me" information which is added to their profile page, and define Preferences (which are prefilled in the special requirements field of the Booking Confirmation form). Their Edit Account page now prompts them to enter such private information as driving license, passport and IBAN number (suggested only, not mandatory).

Private data is viewable by the user themselves, their hosts when they've made a booking at a property, and super managers only.

List Bookings/Invoices/Guests pages offer links to the guest's profiles, and in the List Guests page a receptionist or manager can add a review of the guest, which is visible in the Guest's profile page. This means that when managers/hosts are viewing booking requests they can view both the guest's reviews of other properties, plus other host's reviews of the guest. They can also view the guest's About me section and the private data, all of which should help managers when they're deciding whether or not to approve a booking.

The guest's profile (including private data) are also viewable in the Guest tab of the view booking page.