Added setting which allows property managers to determine if the number of children can be used when checking the number of  guests in a booking v the number of selected rooms.
Enabled/disabled a couple of settings based on an upcoming new "secret" setting that allows some sites to revert to the old guest types if required.

Language files updated with new strings.

If per person per night is not true (i.e. the property/rooms is/are booked per night instead) then child rates are not prompted for, and child prices aren't calculated. Likewise, prices aren't shown for children, but children can still be added to the booking.