Removed container class div from top.html & bottom.html. Where the theme/template allows, then content can be full width. Moved the container instead to wrap the menu position div.
Added a new disable_main_menu script that the main menu will obey, if it has been run then the menu will not be shown anywhere in the page. This allows us to use a simple shortcode {jomres_script disable_main_menu} in any Jomres template and we can disable the menu only on those pages that the template has been rendered.

Added new arguments to show_main_menu that can be used to set which div id the menu should be moved into once rendered. In BS5 and greater, the bottom.html template will trigger the menu, instead of it being rendered elsewhere.