Re-enabled language context switching in Translate lang file strings page. Different property types can have different language labels.
IIRC I disabled this when I removed the automatic creation of language files that were copied when a new property type was created, and this feature was part of that. The multiple language files was confusing users who would update and then not understand why new lang definitions weren't appearing. This was a part of that. Having now checked this out, we can have the feature again without needing the extra language files. <fingers crossed emoji>